# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finax Todo API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh access token",
                "operationId": "abe1b2ca85be2c63cc35c86d3d7e677e",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/AppResponse"
                                        },
                                        {
                                            "properties": {
                                                "auth": {
                                                    "properties": {
                                                        "access_token": {
                                                            "type": "string",
                                                            "example": "1|abc123..."
                                                        },
                                                        "expires_in": {
                                                            "type": "integer",
                                                            "example": 900
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired refresh token"
                    }
                }
            }
        },
        "/api/v1/auth/revoke": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke all tokens",
                "operationId": "4512f41fd65002338c0a6035fc07a480",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/AppResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/check": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Check token validity",
                "operationId": "18199e334fad3647e0c6f3b65b3c1fa2",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/AppResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "user": {
                                                            "$ref": "#/components/schemas/AuthResource"
                                                        },
                                                        "token_expires_at": {
                                                            "type": "string",
                                                            "example": "2026-02-17 12:15:00"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AppResponse": {
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "success",
                            "error",
                            "warning",
                            "info"
                        ],
                        "example": "success"
                    },
                    "toast": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "AuthResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Test User"
                    },
                    "email": {
                        "type": "string",
                        "example": "test@gmail.com"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2026-02-17 12:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}